created a token schema
