+++ api/v1/schemas/role.py
@@ -5,16 +5,3 @@ class RoleCreate(BaseModel):
     role_name: str
     organization_id: str
     permission_ids: List[int]
-
-
-
-# Pydantic models for request and response
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-class TokenData(BaseModel):
-    username: Optional[str] = None
-
-class UserInDB(User):
-    hashed_password: str
